Fix C4242 in GEM font loader
spacing.b_space is std::optional<uint16_t>; glyph_width is int, so the
assignment triggers C4242 on MSVC with /W3 + /WX. Cast to uint16_t
explicitly, matching the style used elsewhere in font_converter.cc.

+++ src/onyx_font/loader/gem_fon.cc
@@ -169,7 +169,7 @@ namespace onyx_font::internal {
 
             // Set spacing (B-space is the character width)
             glyph_spacing spacing;
-            spacing.b_space = glyph_width;
+            spacing.b_space = static_cast<uint16_t>(glyph_width);
             result.m_spacing.push_back(spacing);
 
             if (char_width > 0) {
